% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades_report.r
\docType{data}
\name{average_down_trades_report_analysis_blocks}
\alias{average_down_trades_report_analysis_blocks}
\title{List of Analysis modules used by AverageDownTradesReport Class}
\format{An object of class \code{character} of length 7.}
\usage{
average_down_trades_report_analysis_blocks
}
\description{
List of Analysis modules used by AverageDownTradesReport Class
}
\keyword{datasets}

