% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades_report.r
\docType{class}
\name{AverageDownTradesReport-class}
\alias{AverageDownTradesReport-class}
\title{Average Down Trades Report Data class.}
\description{
Report class computing following blocks:
 "OffsidePositions",
 "OffsidePositionsGainVsDays",
 "OffsidePositionsPnLVsDays",
 "OffsidePositionsCumulativePnL",
 "OffsidePositionsBpsPerMonth",
 "AverageDownTrades",
 "AverageDownTradesFocus"
}
\details{
Also generates summary plot with following subplots:
 "OffsidePositions",
 "OffsidePositionsPnLVsDays",
 "AverageDownTrades",
 "OffsidePositionsCumulativePnL"

Inherits from "VirtualReportAnalysisBlock"
}

