% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades_report.r
\docType{methods}
\name{dataRequest,AverageDownTradesReport,data.frame-method}
\alias{dataRequest,AverageDownTradesReport,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{AverageDownTradesReport,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class 'AverageDownTradesReport'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'AverageDownTradesReport'.
}
\description{
Request data from data source
}

