% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_period_report.r
\docType{methods}
\name{dataRequest,PositionsHoldingPeriodReport,data.frame-method}
\alias{dataRequest,PositionsHoldingPeriodReport,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{PositionsHoldingPeriodReport,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'PositionsHoldingPeriodReport'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'PositionsHoldingPeriodReport'.
}
\description{
Request data from data source
}

