% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_review_report.r
\docType{methods}
\name{dataRequest,RiskReviewReport,data.frame-method}
\alias{dataRequest,RiskReviewReport,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{RiskReviewReport,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class 'RiskReviewReport'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'RiskReviewReport'.
}
\description{
Request data from data source
}

