% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_review_report_core_factors.r
\docType{methods}
\name{dataRequest,RiskReviewReportCoreFactors,data.frame-method}
\alias{dataRequest,RiskReviewReportCoreFactors,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{RiskReviewReportCoreFactors,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'RiskReviewReportCoreFactors'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'RiskReviewReportCoreFactors'.
}
\description{
Request data from data source
}

