% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_tactics_report.r
\docType{methods}
\name{dataRequest,StrategyTacticsReport,data.frame-method}
\alias{dataRequest,StrategyTacticsReport,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{StrategyTacticsReport,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class 'StrategyTacticsReport'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'StrategyTacticsReport'.
}
\description{
Request data from data source
}

