% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_review_2016.2_report.r
\docType{methods}
\name{dataRequest,TradingReview2016.2Report,data.frame-method}
\alias{dataRequest,TradingReview2016.2Report,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{TradingReview2016.2Report,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class 'TradingReview2016.2Report'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'TradingReview2016.2Report'.
}
\description{
Request data from data source
}

