% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trades_report.r
\docType{data}
\name{extended_trades_report_analysis_blocks}
\alias{extended_trades_report_analysis_blocks}
\title{List of Analysis modules used by ExtendedTradesReport Class}
\format{An object of class \code{character} of length 5.}
\usage{
extended_trades_report_analysis_blocks
}
\description{
List of Analysis modules used by ExtendedTradesReport Class
}
\keyword{datasets}

