% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trades_report.r
\docType{class}
\name{ExtendedTradesReport-class}
\alias{ExtendedTradesReport-class}
\title{Extended Trades Report class.}
\description{
Report class computing following blocks:
 "ExtendedTrades",
 "BuysAndSells",
 "MarketReturn",
 "RelativeMarketReturn",
 "ExtendedTradesSummary"
}
\details{
Inherits from "VirtualReportAnalysisBlock"
}

