% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis_block.r
\docType{methods}
\name{getOutputFrontendDataList}
\alias{getOutputFrontendDataList}
\alias{getOutputFrontendDataList,VirtualReportAnalysisBlock-method}
\title{Returns generated output frontend data list}
\usage{
getOutputFrontendDataList(object)

\S4method{getOutputFrontendDataList}{VirtualReportAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualReportAnalysisBlock"}
}
\value{
\code{frontend_data_list} list of objects of class "data.frame"
}
\description{
Returns list of data to be used by any frontend using the module
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualReportAnalysisBlock}: Returns generated output frontend data list

Returns list of data to be used by any frontend using the module
}}

