% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis_block.r
\docType{methods}
\name{getOutputGGPlotDataList}
\alias{getOutputGGPlotDataList}
\alias{getOutputGGPlotDataList,VirtualReportAnalysisBlock-method}
\title{Returns list of generated output ggplot data.frame's}
\usage{
getOutputGGPlotDataList(object)

\S4method{getOutputGGPlotDataList}{VirtualReportAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualReportAnalysisBlock"}
}
\value{
\code{ggplot_data_list} list of objects of class "data.frame"
}
\description{
Returns list of "data.frame"'s with data used to generate ggplots
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualReportAnalysisBlock}: Returns list of generated output ggplot data.frame's

Returns list of "data.frame"'s with data used to generate ggplots
}}

