% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis_block.r
\docType{methods}
\name{getOutputGGPlotList}
\alias{getOutputGGPlotList}
\alias{getOutputGGPlotList,VirtualReportAnalysisBlock-method}
\title{Returns generated output ggplot list}
\usage{
getOutputGGPlotList(object)

\S4method{getOutputGGPlotList}{VirtualReportAnalysisBlock}(object)
}
\arguments{
\item{object}{object of class "VirtualReportAnalysisBlock"}
}
\value{
\code{ggplot_list} list of objects of class "ggplot" or "grob"
}
\description{
Returns list of ggplot or grob objects generated
as a result of analysis computation.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualReportAnalysisBlock}: Returns generated output ggplot list

Returns list of ggplot or grob objects generated as a result of analysis computation.
}}

