% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_period_report.r
\docType{data}
\name{positions_holding_period_report_analysis_blocks}
\alias{positions_holding_period_report_analysis_blocks}
\title{List of Analysis modules used by PositionsHoldingPeriodReport Class}
\format{An object of class \code{character} of length 3.}
\usage{
positions_holding_period_report_analysis_blocks
}
\description{
List of Analysis modules used by PositionsHoldingPeriodReport Class
}
\keyword{datasets}

