% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_period_report.r
\docType{class}
\name{PositionsHoldingPeriodReport-class}
\alias{PositionsHoldingPeriodReport-class}
\title{Extended Trades Report class.}
\description{
Report class computing following blocks:
 "PositionsHoldingPeriod",
 "PositionsHoldingDayZeroPnL",
 "PositionsHoldingCapitalDistribution"
}
\details{
Inherits from "VirtualReportAnalysisBlock"
}

