% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis_block.r
\name{.copyAnalyzerOutputData}
\alias{.copyAnalyzerOutputData}
\title{Copy output of analyzer to internal list}
\usage{
.copyAnalyzerOutputData(object, analyzer)
}
\arguments{
\item{object}{object of class "VirtualReportAnalysisBlock"}

\item{analyzer}{object of type "VirtualAnalysisBlock"}
}
\value{
\code{object} object object of class "VirtualReportAnalysisBlock"
}
\description{
Private method to copy analyzer output and deal to internal lists
}

