% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis_block.r
\name{.setOutputFrontendDataList}
\alias{.setOutputFrontendDataList}
\title{Set internaly list of data for ggplot objects}
\usage{
.setOutputFrontendDataList(object, list)
}
\arguments{
\item{object}{object of class "VirtualReportAnalysisBlock"}

\item{list}{list of objects of class "data.frame"}
}
\value{
\code{object} object object of class "VirtualReportAnalysisBlock"
}
\description{
Private method to set frontend_data_list slot
with list of data.frame's used to provide to
frontend services object
}

