% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis_block.r
\name{.setOutputGGPlotDataList}
\alias{.setOutputGGPlotDataList}
\title{Set internaly list of data for ggplot objects}
\usage{
.setOutputGGPlotDataList(object, list)
}
\arguments{
\item{object}{object of class "VirtualReportAnalysisBlock"}

\item{list}{list of objects of class "data.frame"}
}
\value{
object object of class "VirtualReportAnalysisBlock"
}
\description{
Private method to set slot with list of data
used to generate ggplot object.
}

