% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis_block.r
\name{.setOutputGGPlotList}
\alias{.setOutputGGPlotList}
\title{Set internaly output ggplot object list}
\usage{
.setOutputGGPlotList(object, list)
}
\arguments{
\item{object}{object of class "VirtualReportAnalysisBlock"}

\item{list}{list with objects of class "ggplot"}
}
\value{
\code{object} object object of class "VirtualReportAnalysisBlock"
}
\description{
Private method to set ggplot slot with generated GGPlot object list
}

