% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis_block.r
\name{.setOutputObjectList}
\alias{.setOutputObjectList}
\title{Set internaly output object list}
\usage{
.setOutputObjectList(object, list)
}
\arguments{
\item{object}{object of class "VirtualReportAnalysisBlock"}

\item{list}{list of objects of class derived from "VirtualReferenceData"}
}
\value{
\code{object} object object of class "VirtualReportAnalysisBlock"
}
\description{
Private method to set output object list slot
}

