% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_down_trades_report.r
\docType{methods}
\name{Process,AverageDownTradesReport-method}
\alias{Process,AverageDownTradesReport-method}
\title{Trigger computation of report data.}
\usage{
\S4method{Process}{AverageDownTradesReport}(object)
}
\arguments{
\item{object}{object of class "AverageDownTradesReport"}
}
\value{
\code{object} object object of class "AverageDownTradesReport"
}
\description{
Trigger computation of report data.
}

