% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trades_report.r
\docType{methods}
\name{Process,ExtendedTradesReport-method}
\alias{Process,ExtendedTradesReport-method}
\title{Trigger computation of report data.}
\usage{
\S4method{Process}{ExtendedTradesReport}(object)
}
\arguments{
\item{object}{object of class "ExtendedTradesReport"}
}
\value{
\code{object} object object of class "ExtendedTradesReport"
}
\description{
Trigger computation of report data.
}

