% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_holding_period_report.r
\docType{methods}
\name{Process,PositionsHoldingPeriodReport-method}
\alias{Process,PositionsHoldingPeriodReport-method}
\title{Trigger computation of report data.}
\usage{
\S4method{Process}{PositionsHoldingPeriodReport}(object)
}
\arguments{
\item{object}{object of class "PositionsHoldingPeriodReport"}
}
\value{
\code{object} object object of class "PositionsHoldingPeriodReport"
}
\description{
Trigger computation of report data.
}

