% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_review_report.r
\docType{methods}
\name{Process,RiskReviewReport-method}
\alias{Process,RiskReviewReport-method}
\title{Trigger computation of report data.}
\usage{
\S4method{Process}{RiskReviewReport}(object)
}
\arguments{
\item{object}{object of class "PositionsHoldingPeriodReport"}
}
\value{
\code{object} object object of class "PositionsHoldingPeriodReport"
}
\description{
Trigger computation of report data.
}

