% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_review_report_core_factors.r
\docType{methods}
\name{Process,RiskReviewReportCoreFactors-method}
\alias{Process,RiskReviewReportCoreFactors-method}
\title{Trigger computation of report data.}
\usage{
\S4method{Process}{RiskReviewReportCoreFactors}(object)
}
\arguments{
\item{object}{object of class "PositionsHoldingPeriodReport"}
}
\value{
\code{object} object object of class "PositionsHoldingPeriodReport"
}
\description{
Trigger computation of report data.
}

