% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_tactics_report.r
\docType{methods}
\name{Process,StrategyTacticsReport-method}
\alias{Process,StrategyTacticsReport-method}
\title{Trigger computation of report data.}
\usage{
\S4method{Process}{StrategyTacticsReport}(object)
}
\arguments{
\item{object}{object of class "StrategyTacticsReport"}
}
\value{
\code{object} object object of class "StrategyTacticsReport"
}
\description{
Trigger computation of report data.
}

