% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_review_2016.2_report.r
\docType{methods}
\name{Process,TradingReview2016.2Report-method}
\alias{Process,TradingReview2016.2Report-method}
\title{Trigger computation of report data.}
\usage{
\S4method{Process}{TradingReview2016.2Report}(object)
}
\arguments{
\item{object}{object of class "TradingReview2016.2Report"}
}
\value{
\code{object} object object of class "TradingReview2016.2Report"
}
\description{
Trigger computation of report data.
}

