% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_review_report.r, R/risk_review_report_core_factors.r
\docType{data}
\name{risk_review_analysis_blocks}
\alias{risk_review_analysis_blocks}
\title{List of Analysis modules used by RiskReviewReport Class}
\format{An object of class \code{character} of length 7.}
\usage{
risk_review_analysis_blocks

risk_review_analysis_blocks
}
\description{
List of Analysis modules used by RiskReviewReport Class

List of Analysis modules used by RiskReviewReportCoreFactors Class
}
\keyword{datasets}

