% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_review_report.r
\docType{class}
\name{RiskReviewReport-class}
\alias{RiskReviewReport-class}
\title{Risk Review Report class.}
\description{
Report class computing following blocks:
 "PortfolioVarianceDecomposition",
 "PortfolioFactorReturns",
 "PortfolioFactorExposures"
}
\details{
Inherits from "VirtualReportAnalysisBlock"
}

