% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_review_report_core_factors.r
\docType{class}
\name{RiskReviewReportCoreFactors-class}
\alias{RiskReviewReportCoreFactors-class}
\title{Risk Review Report class.}
\description{
Report class computing following blocks:
 "PortfolioVarianceDecomposition",
 "PortfolioFactorReturns",
 "PortfolioFactorExposures"
}
\details{
Inherits from "VirtualReportAnalysisBlock"
}

