% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_tactics_report.r
\docType{data}
\name{strategy_tactics_report_analysis_blocks}
\alias{strategy_tactics_report_analysis_blocks}
\title{List of Analysis modules used by StrategyTacticsReport Class}
\format{An object of class \code{character} of length 10.}
\usage{
strategy_tactics_report_analysis_blocks
}
\description{
List of Analysis modules used by StrategyTacticsReport Class
}
\keyword{datasets}

