% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_tactics_report.r
\docType{class}
\name{StrategyTacticsReport-class}
\alias{StrategyTacticsReport-class}
\title{Strategy Tactics Report class.}
\description{
Report class computing following blocks:
 "StrategyBreakdown",
 "StrategyBreakdownAUMAndTurnover",
 "StrategyBreakdownTotalAndDeltaPnL",
 "StrategyBreakdownValueTradedPerSignal",
 "StrategyBreakdownPnLOnTradeDayPerSignal",
 "StrategyBreakdownPnLAndTurnoverPerEvent",
 "StrategyBreakdownSignalCharacteristicAndEffectiveness",
 "PortfolioFactorExposure",
 "ValueTradedInLongShortHedge",
 "PnLTradedInLongShortHedge"
}
\details{
Also generates summary plot with following subplots:
 "StrategyBreakdownAUMAndTurnover",
 "StrategyBreakdownTotalAndDeltaPnL"
 "StrategyBreakdownPnLAndTurnoverPerEvent",
 "StrategyBreakdownSignalCharacteristicAndEffectiveness",
 "PortfolioFactorExposure"

Inherits from "VirtualReportAnalysisBlock"
}

