% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_review_2016.2_report.r
\docType{data}
\name{trading_review_2016.2_report_analysis_blocks}
\alias{trading_review_2016.2_report_analysis_blocks}
\title{List of Analysis modules used by TradingReview2016.2Report Class}
\format{An object of class \code{character} of length 10.}
\usage{
trading_review_2016.2_report_analysis_blocks
}
\description{
List of Analysis modules used by TradingReview2016.2Report Class
}
\keyword{datasets}

