% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_review_2016.2_report.r
\docType{class}
\name{TradingReview2016.2Report-class}
\alias{TradingReview2016.2Report-class}
\title{Strategy Tactics Report class.}
\description{
Report class computing following blocks:
 "OffsidePositions",
 "OffsidePositionsBpsPerMonth",
 "AverageDownTradesFocus",
 "PositionsHoldingDayZeroPnL",
 "PositionsHoldingCapitalDistribution",
 "StrategyBreakdownAUMAndTurnover",
 "StrategyBreakdownTotalAndDeltaPnL",
 "StrategyBreakdownPnLAndTurnoverPerEvent",
 "PortfolioFactorExposure",
 "PositionRevisitsDeltaPrevQuarter"
}
\details{
Also generates summary plot with following subplots:
 "StrategyBreakdownAUMAndTurnover",
 "StrategyBreakdownTotalAndDeltaPnL"
 "OffsidePositions",
 "PositionsHoldingCapitalDistribution"
 "PositionRevisitsDeltaPrevQuarter"
 "StrategyBreakdownPnLAndTurnoverPerEvent",
 "PortfolioFactorExposure"

Inherits from "VirtualReportAnalysisBlock"
}

