% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_analysis_block.r
\docType{class}
\name{VirtualReportAnalysisBlock-class}
\alias{VirtualReportAnalysisBlock-class}
\title{Virtual S4 class implementing handling of analysis objects.}
\description{
Class that is to be inherited by any objects
that will encapsulate specific analysis computations.
}
\details{
Each analysis module will generate an entry to the list
with respective class name.

Inherits from "VirtualDataSourceClient"
}
\section{Slots}{

\describe{
\item{\code{ggplot_list}}{list of objects of class "ggplot" or "grob"}

\item{\code{ggplot_data_list}}{list of "data.frame"'s with data for ggplots}

\item{\code{frontend_data}}{list of "data.frame"'s with data for ggplots}

\item{\code{output}}{list of output from analysis classes of class derived from "VirtualReferenceData"}
}}

